@@ -8,7 +8,9 @@
         "open graph",
         "opengraph",
         "twitter image",
-        "vk image"
+        "vk image",
+        "Laravel 11",
+        "Laravel 12"
     ],
     "authors": [
         {
@@ -19,19 +21,19 @@
     ],
     "homepage": "https://github.com/abordage/laravel-og-images",
     "require": {
-        "php": ">=7.4",
+        "php": "^8.2",
         "ext-imagick": "*",
         "abordage/og-images": "^0.4.1",
-        "illuminate/support": "^8.0 || ^9.0 || ^10.0 || ^11.0"
+        "illuminate/support": "^11.0 || ^12.0"
     },
     "require-dev": {
         "ext-fileinfo": "*",
         "friendsofphp/php-cs-fixer": "^3.0",
+        "larastan/larastan": "^2.9 || ^3.0",
         "mikey179/vfsstream": "^1.6",
-        "nunomaduro/collision": "^5.0 || ^6.0 || ^7.0 || ^8.0",
-        "nunomaduro/larastan": "^1.0 || ^2.0",
-        "orchestra/testbench": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "phpunit/phpunit": "^9.6 || ^10.0"
+        "nunomaduro/collision": "^8.0",
+        "orchestra/testbench": "^9.0 || ^10.0",
+        "phpunit/phpunit": "^10.0 || ^11.0"
     },
     "minimum-stability": "stable",
     "prefer-stable": true,
